print the list safe

+++ 0x13-more_singly_linked_lists/101-print_listint_safe.c
@@ -10,7 +10,7 @@
 size_t print_listint_safe(const listint_t *head)
 {
 const listint_t *position, *next_p = head;
-size_t cont = 0;
+unsigned int cont = 0;
 if (head == '\0')
 return ('\0');
 while (next_p != 0)
